feat(admin): move certif scoring simulator in tools
