+++ admin/app/components/tools/certification.gjs
@@ -1,0 +1,3 @@
+import ScoringSimulator from './certification/scoring-simulator';
+
+<template><ScoringSimulator /></template>
