+++ admin/app/templates/authenticated/tools.hbs
@@ -10,6 +10,9 @@
       <LinkTo @route="authenticated.tools.campaigns">
         {{t "pages.administration.navigation.campaigns.label"}}
       </LinkTo>
+      <LinkTo @route="authenticated.tools.certification">
+        {{t "pages.administration.navigation.certification.label"}}
+      </LinkTo>
       <LinkTo @route="authenticated.tools.junior">
         {{t "pages.administration.navigation.junior.label"}}
       </LinkTo>
