+++ admin/app/templates/authenticated/tools/certification.hbs
@@ -1,0 +1,1 @@
+<Tools::Certification />
